+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - PixPullRequest resource
 - IssuingStockRule resource
 - rules and debtorWorkspaceId parameters to CreditNote resource
+- reason attribute to PixRequest resource
 
 ## [0.18.0] - 2026-06-22
 ### Changed